# contracts/http/openapi.yaml
openapi: 3.0.3
info:
  title: Nadir API
  version: 1.0.0
  description: JSON and Server-Sent Events API for document-grounded search.
servers:
  - url: /
paths:
...
  /api/v1/turns/{id}/events:
    get:
      operationId: streamTurn
      parameters:
        - { $ref: '#/components/parameters/TurnID' }
        - in: header
          name: Last-Event-ID
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Ordered turn events.
          content:
            text/event-stream: {}
        '204': { description: Stream already fully replayed. }
  /api/v1/turns/{id}/cancel:
    post:
      operationId: cancelTurn
      parameters: [{ $ref: '#/components/parameters/TurnID' }]
      responses:
        '204': { description: Cancellation accepted. }
...
components:
  parameters:
    TurnID:
      name: id
      in: path
      required: true
      schema: { type: string }
...